refactor: Move ProgramsList-component to page-only scope
