+++ interfaces/Portal/src/app/pages/home/components/programs-list/programs-list.component.scss
